# docs/source/specs/openapi.yml
swagger: '2.0'
info:
  description: 'The API for Role Based Access Control.'
  version: '1.0.0'
  title: 'Role Based Access Control'
  license:
    name: 'AGPL-3.0'
    url: 'https://opensource.org/licenses/AGPL-3.0'
basePath: '/r/insights/platform/rbac/api/v1'
...
schemes:
- 'http'
paths:
...
  /groups/:
    post:
      tags:
        - Group
      summary: Create a group in a tenant
      operationId: createGroup
      consumes:
      - 'application/json'
      produces:
      - 'application/json'
      parameters:
      - in: 'body'
        name: 'body'
        description: Group to create in tenant
        required: true
        schema:
          $ref: '#/definitions/Group'
      responses:
        '201':
          description: An object describing the group
          schema:
            $ref: '#/definitions/GroupOut'
        '401':
          description: Unauthorized
        '500':
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - api_key: []
    get:
      tags:
        - Group
      summary: List the groups for a tenant
      operationId: listGroups
      produces:
      - 'application/json'
      responses:
        '200':
          description: A paginated list of group objects
          schema:
            $ref: '#/definitions/GroupPagination'
        '401':
          description: Unauthorized
        '500':
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - api_key: []
  /groups/{uuid}/:
    get:
      tags:
        - Group
      summary: Get a group in the tenant
      operationId: getGroup
      produces:
      - 'application/json'
      parameters:
      - name: 'uuid'
        in: 'path'
        description: 'ID of group to get'
        required: true
        type: 'string'
        format: 'uuid'
      responses:
        '200':
          description: A Group object
          schema:
            $ref: '#/definitions/GroupWithPrincipals'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - api_key: []
    delete:
      tags:
        - Group
      summary: Delete a group in the tenant
      operationId: deleteGroup
      parameters:
      - name: 'uuid'
        in: 'path'
        description: 'ID of group to delete'
        required: true
        type: 'string'
        format: 'uuid'
      responses:
        '204':
          description: Group deleted
        '401':
          description: Unauthorized
        '403':
          description: Insufficent permissions to delete group
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - api_key: []
...
securityDefinitions:
  api_key:
    type: 'apiKey'
    name: 'Authorization'
    in: 'header'
definitions:
  Error:
    required:
      - error_key
    properties:
      error_key:
        type: string
  ListPagination:
    properties:
      count:
        type: integer
        format: int64
        example: 30
      first:
        type: string
        format: uri
        example: /api/v1/(resources)/?page=1
      previous:
        type: string
        format: uri
        example: /api/v1/(resources)/?page=2
      next:
        type: string
        format: uri
        example: /api/v1/(resources)/?page=4
      last:
        type: string
        format: uri
        example: /api/v1/(resources)/?page=10
  Principal:
    required:
      - username
      - email
    properties:
      username:
        type: string
        example: smithj
      email:
        type: string
        format: email
        example: smithj@mytechco.com
      first_name:
        type: string
        example: John
      last_name:
        type: string
        example: Smith
      is_org_admin:
        type: boolean
        example: true
...
  PrincipalOut:
    allOf:
      - $ref: '#/definitions/Principal'
    required:
      - uuid
    properties:
      uuid:
        type: string
        format: uuid
        example: 57e60f90-8c0c-4bd1-87a0-2143759aae1c
...
  Group:
    required:
      - name
    properties:
      name:
        type: string
        example: GroupA
  GroupOut:
    allOf:
      - $ref: '#/definitions/Group'
    required:
      - uuid
    properties:
      uuid:
        type: string
        format: uuid
        example: 57e60f90-8c0c-4bd1-87a0-2143759aae1c
  GroupWithPrincipals:
    allOf:
      - $ref: '#/definitions/GroupOut'
    required:
      - principals
    properties:
      uuid:
        type: string
        format: uuid
        example: 57e60f90-8c0c-4bd1-87a0-2143759aae1c
      principals:
        type: array
        items:
          $ref: '#/definitions/PrincipalOut'
  GroupPagination:
    allOf:
      - $ref: '#/definitions/ListPagination'
    required:
      - data
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/GroupOut'
...